requirements.txt
amqp==2.5.2
appdirs==1.4.3
arrow==0.15.5
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
automium==0.2.6
automium-web==0.1.1
base==1.0.4
beautifulsoup4==4.8.1
billiard==3.6.3.0
camera==1.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
constants==0.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
dateparser==0.7.4
decorator==4.4.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-classy-tags==0.9.0
django-cms==3.7.0
django-crispy-forms==1.9.0
django-datetime-widget==0.9.3
django-filer==1.7.0
django-formtools==2.1
django-hero-slider==1.3.1
django-hvad==1.8.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.0.0
django-select2==7.2.3
django-tempus-dominus==5.1.2.13
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangorestframework==3.11.0
docopt==0.6.2
dominate==2.5.1
drf-yasg==1.17.1
easy-thumbnails==2.7
ExtensionClass==4.4
filelock==3.0.12
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flower==0.9.4
future==0.18.1
gunicorn==20.0.4
humanize==0.5.1
hupper==1.9
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
kombu==4.6.8
legacy==0.1.6
MarkupSafe==1.1.1
moment==0.8.2
more-itertools==7.2.0
networkx==2.4
nltk==3.4.5
numpy==1.18.1
packaging==20.3
PasteDeploy==2.0.1
pika==0.9.5
pilkit==2.0
Pillow==7.1.1
plaster==1.0
plaster-pastedeploy==0.7
player==0.6.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.0
pygame==1.9.6
pyparsing==2.4.7
pyramid==1.10.4
pyramid-jinja2==2.8
pytest==5.3.1
python-dateutil==2.8.0
pythonanywhere==0.7.4
pytz==2019.3
Record==3.5
redis==3.5.1
regex==2020.1.8
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scapy==2.4.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
slugify==0.0.1
soupsieve==1.9.5
SQLAlchemy==1.3.15
sqlparse==0.3.1
tiles==1.0.4
times==0.7
tornado==6.0.4
translationstring==1.3
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
venusian==1.2.0
view==0.1
vine==1.3.0
virtualenv==20.0.16
visitor==0.1.3
wcwidth==0.1.7
WebOb==1.8.5
Werkzeug==0.15.2
whitenoise==5.0.1
WTForms==2.2.1
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==4.6.0
